Main gets a driver.
